Created utility py script, added basic snippets to tex editing.
Python script duplicates a .lua file and removes the comments, because I like having the educational comments in my init.lua; however, somtimes it can be very difficult to read.

Added basic snippets functionality, such as moving through nodes,
repeated typing, and some regex.
In the future, I need to replace some of the auto commands that use
regex with "{ condition = tex.in_mathzone }."
After that, I need to look up a list of snippets and copy the ones that
I like.